backend/requirements.txt
amqp==5.3.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
billiard==4.2.4
celery==5.6.0
cffi==2.0.0
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
croniter==6.0.0
cryptography==46.0.3
dnspython==2.8.0
ecdsa==0.19.1
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.124.4
ffmpeg-python==0.2.0
future==1.0.0
greenlet==3.3.0
h11==0.16.0
idna==3.11
kombu==5.6.1
packaging==25.0
passlib==1.7.4
prompt_toolkit==3.0.52
pyasn1==0.6.1
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.8.2
pydantic_core==2.41.5
python-dateutil==2.9.0.post0
python-jose[cryptography]==3.3.0
python-multipart==0.0.20
pytz==2025.2
redis==6.4.0
rq==2.6.1
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.45
sqlmodel==0.0.27
starlette==0.50.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uvicorn==0.38.0
vine==5.1.0
wcwidth==0.2.14
slowapi==0.1.9
assemblyai==0.37.0
pytest==8.3.5
httpx==0.28.1
boto3==1.35.11
botocore==1.35.11
pyotp==2.9.0
sentry-sdk[fastapi]==2.14.0
structlog==23.3.0
python-json-logger==2.0.7
prometheus-client==0.21.0
psutil==6.1.0
bleach==6.2.0
qrcode[pil]==7.4.2
requests==2.32.3
stripe==10.12.0
google-cloud-storage==2.18.2
